Model: Stores vertex's correctly

+++ vec.h
@@ -1,9 +1,12 @@
 struct vec2 {
-    int x, y;
+    //represents a mathmatical 2D vector. Contains x and y axis
+    float x, y;
+    
 
 };
 
 struct vec3 {
-    int x, y, z;
+    //represents a mathmatical 3D vector. Contains x, y, z axis
+    float x, y, z;
 
 };
